Use long int instead of int for nasl.
Similar to what's done in Master branch.

+++ nasl/exec.h
@@ -20,6 +20,6 @@
 #define EXEC_H_INCLUDED
 
 tree_cell *nasl_exec (lex_ctxt *, tree_cell *);
-int cell_cmp (lex_ctxt *, tree_cell *, tree_cell *);
+long int cell_cmp (lex_ctxt *, tree_cell *, tree_cell *);
 tree_cell *cell2atom (lex_ctxt *, tree_cell *);
 #endif
